requirements.txt.txt
# Core Libraries
python-dotenv==1.0.0
gradio==4.13.0

# Speech Processing
edge-tts==6.1.3
pygame==2.5.2  # For audio playback

# STT & Browser Control
selenium==4.19.0
webdriver-manager==4.0.1

# LLM & NLP
langchain-groq==0.1.2
langchain-google-genai==0.0.11
langchain-chroma==0.1.1
langchain-huggingface==0.0.21
langchain==0.1.13

# Embeddings & VectorDB
chromadb==0.4.24
sentence-transformers==2.7.0

# Arabic NLP Specific
mtranslate==1.8.2  # Fallback translation
transformers[torch]==4.41.0  # For AraBERT

# Email & Crisis Alerts
smtplib-starttls==1.0.0

# Utility
tqdm==4.66.2  # Progress bars
python-multipart==0.0.6  # File handling